requirements.txt
streamlit==1.39.0
tensorflow==2.16.0
pillow==10.2.0
numpy==1.26.4